+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/network/model/response/ChoreResponseBase.kt
@@ -1,4 +1,4 @@
-package api.model.response
+package com.altmann.choresmanager.network.model.response
 
 import androidx.compose.ui.graphics.Color
 import com.altmann.choresmanager.models.chores.Priority
@@ -10,7 +10,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 sealed interface ChoreResponseBase {
-    val choreId: Int
+    val choreId: Int?
     @Contextual
     val startTime: LocalTime
     @Contextual
@@ -27,8 +27,5 @@ sealed interface ChoreResponseBase {
     val deleted: Boolean?
     var finished: Boolean
     @Contextual
-    val finishedDate: LocalDate?
-    @Contextual
     val color: Color
-    val userId: Int
 }