+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/network/model/response/CollegeChoreResponse.kt
@@ -1,4 +1,4 @@
-package api.model.response
+package com.altmann.choresmanager.network.model.response
 
 import androidx.compose.ui.graphics.Color
 import com.altmann.choresmanager.models.chores.Priority
@@ -12,7 +12,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 @SerialName("COLLEGE")
 data class CollegeChoreResponse(
-    override val choreId: Int,
+    override val choreId: Int?,
     @Contextual
     override val startTime: LocalTime,
     @Contextual
@@ -24,15 +24,12 @@ data class CollegeChoreResponse(
     override val endDate: LocalDate,
     override var choreException: List<@Contextual LocalDate> = emptyList(),
     override val title: String,
-    override val description: String?,
+    override val description: String? = null,
     override var priority: Priority,
     override val deleted: Boolean? = false,
     override var finished: Boolean = false,
     @Contextual
-    override val finishedDate: LocalDate?,
-    @Contextual
     override val color: Color,
-    override val userId: Int,
 
     val subject : String,
     val location: String?,