+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/repository/ChoreRepository.kt
@@ -1,8 +1,10 @@
 package com.altmann.choresmanager.repository
 
-import api.model.response.ChoreResponseBase
+import com.altmann.choresmanager.models.chores.Chore
+import com.altmann.choresmanager.network.model.response.ChoreResponseBase
 import com.altmann.choresmanager.network.ApiClient
 import com.altmann.choresmanager.network.ApiResult
+import com.altmann.choresmanager.utils.ResponseToChore
 
 class ChoreRepository(private val api: ApiClient = ApiClient()) : IChoreRepository{
 
@@ -10,6 +12,13 @@ class ChoreRepository(private val api: ApiClient = ApiClient()) : IChoreReposito
 
     override suspend fun getChores(userId: Int): ApiResult<List<ChoreResponseBase>> {
         return api.get("$URL$userId")
+    }
+
+    override suspend fun addChore(
+        userId: Int,
+        chore: Chore
+    ): ApiResult<ChoreResponseBase> {
+        return api.post("$URL$userId", ResponseToChore.toResponse(chore))
     }
 
 }