+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/repository/IChoreRepository.kt
@@ -1,8 +1,10 @@
 package com.altmann.choresmanager.repository
 
-import api.model.response.ChoreResponseBase
+import com.altmann.choresmanager.models.chores.Chore
+import com.altmann.choresmanager.network.model.response.ChoreResponseBase
 import com.altmann.choresmanager.network.ApiResult
 
 interface IChoreRepository {
     suspend fun getChores(userId : Int) : ApiResult<List<ChoreResponseBase>>
+    suspend fun addChore(userId: Int, chore: Chore) : ApiResult<ChoreResponseBase>
 }